Tools/requirements.txt
pycryptodome
pycryptodomex
ecdsa
pyelftools
intelhex
